+++ wemake_python_styleguide/violations/consistency.py
@@ -1403,7 +1403,7 @@ class InconsistentReturnVariableViolation(ASTViolation):
 @final
 class WalrusViolation(ASTViolation):
     """
-    Forbid walrus operator.
+    Forbid the use of the walrus operator (`:=`) outside of comprehensions.
 
     Reasoning:
         Code with ``:=`` is hardly readable.
@@ -1412,7 +1412,8 @@ class WalrusViolation(ASTViolation):
         Python is not expression-based.
 
     Solution:
-        Don't use fancy stuff, use good old assignments.
+        Avoid using the walrus operator outside comprehensions.
+        Stick to traditional assignment statements for clarity.
 
     Example::
 
@@ -1429,7 +1430,7 @@ class WalrusViolation(ASTViolation):
 
     """
 
-    error_template = 'Found walrus operator'
+    error_template = 'Found walrus operator outside a comprehension'
     code = 332