+++ wemake_python_styleguide/visitors/ast/operators.py
@@ -191,7 +191,8 @@ class WrongMathOperatorVisitor(base.BaseNodeVisitor):
 
     def _check_list_multiply(self, node: ast.BinOp) -> None:
         is_list_multiply = isinstance(node.op, ast.Mult) and isinstance(
-            node.left, self._list_nodes
+            node.left,
+            self._list_nodes,
         )
         if is_list_multiply:
             self.add_violation(ListMultiplyViolation(node.left))
@@ -223,10 +224,27 @@ class WrongMathOperatorVisitor(base.BaseNodeVisitor):
 class WalrusVisitor(base.BaseNodeVisitor):
     """We use this visitor to find walrus operators and ban them."""
 
+    _available_parents: ClassVar[AnyNodes] = (
+        ast.ListComp,
+        ast.SetComp,
+        ast.DictComp,
+        ast.GeneratorExp,
+    )
+
     def visit_NamedExpr(
         self,
         node: ast.NamedExpr,
     ) -> None:
-        """Disallows walrus ``:=`` operator."""
-        self.add_violation(consistency.WalrusViolation(node))
+        """Disallows walrus ``:=`` operator outside comprehensions."""
+        self._check_walrus_in_comprehesion(node)
         self.generic_visit(node)
+
+    def _check_walrus_in_comprehesion(
+        self,
+        node: ast.NamedExpr,
+    ) -> None:
+        is_comprension = walk.get_closest_parent(node, self._available_parents)
+        if is_comprension:
+            return
+
+        self.add_violation(consistency.WalrusViolation(node))